src/services/scrapers/base_scraper.py
"""
Scraper de base pour VIGIL

Placeholder pour les scrapers avancés (Phase 2)
"""

import asyncio
from abc import ABC, abstractmethod
from typing import Optional, Dict, Any
import httpx
from bs4 import BeautifulSoup
from loguru import logger

from ...config import settings


class BaseScraper(ABC):
    """
    Scraper de base avec fonctionnalités communes :
    - Rate limiting
    - Retry automatique
    - Respect des robots.txt
    - User-Agent rotation
    """

    def __init__(self, base_url: str):
        self.base_url = base_url.rstrip("/")
        self.client: Optional[httpx.AsyncClient] = None

    async def __aenter__(self):
        """Context manager entry"""
        self.client = httpx.AsyncClient(
            timeout=30.0,
            headers={"User-Agent": settings.USER_AGENT},
            follow_redirects=True
        )
        return self

    async def __aexit__(self, exc_type, exc_val, exc_tb):
        """Context manager exit"""
        if self.client:
            await self.client.aclose()

    async def scrape(self, url: str) -> Optional[BeautifulSoup]:
        """
        Scrape une URL et retourne le soup BeautifulSoup.

        Args:
            url: URL à scraper

        Returns:
            BeautifulSoup object ou None
        """
        if not self.client:
            raise RuntimeError("Client not initialized. Use 'async with' context manager.")

        try:
            # Rate limiting
            await asyncio.sleep(settings.SCRAPING_DELAY)

            # Requête
            logger.debug(f"Scraping {url}")
            response = await self.client.get(url)

            if response.status_code != 200:
                logger.warning(f"Failed to scrape {url}: {response.status_code}")
                return None

            # Parser
            soup = BeautifulSoup(response.text, 'html.parser')
            return soup

        except Exception as e:
            logger.error(f"Error scraping {url}: {e}")
            return None

    @abstractmethod
    async def extract_data(self, url: str) -> Optional[Dict[str, Any]]:
        """
        Extrait les données d'une URL.
        À implémenter par les sous-classes.
        """
        pass
