requirements.txt
# VIGIL - Expert Concurrence Module
# Requirements for Phase 1 (POC)

# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & HTTP
asyncio==3.4.3
aiohttp==3.9.1
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
scrapy==2.11.0
playwright==1.40.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# APIs clients
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
